CodePushUtils moved to shared project
